refactor: extract Node creation logic out
bahevior preserving refactoring before big feature addition
