+++ lib/assertion-visitor.js
@@ -152,19 +152,6 @@ const isPathIdentical = (path1, path2) => {
   return path1.join('/') === path2.join('/');
 };
 
-const findBlockedScope = (scopeStack) => {
-  const lastIndex = scopeStack.length - 1;
-  const scope = scopeStack[lastIndex];
-  if (!scope.block || isArrowFunctionWithConciseBody(scope.block)) {
-    return findBlockedScope(scopeStack.slice(0, lastIndex));
-  }
-  return scope;
-};
-
-const isArrowFunctionWithConciseBody = (node) => {
-  return node.type === 'ArrowFunctionExpression' && node.body.type !== 'BlockStatement';
-};
-
 const isFunction = (node) => {
   switch (node.type) {
     case syntax.FunctionDeclaration:
